+++ src/onegov.election_day/onegov/election_day/views/manage/elections.py
@@ -8,7 +8,7 @@ from onegov.election_day import _
 from onegov.election_day import ElectionDayApp
 from onegov.election_day.collections import ArchivedResultCollection
 from onegov.election_day.collections import NotificationCollection
-from onegov.election_day.forms import DeleteForm
+from onegov.election_day.forms import EmptyForm
 from onegov.election_day.forms import ElectionForm
 from onegov.election_day.forms import TriggerNotificationForm
 from onegov.election_day.layout import ManageElectionsLayout
@@ -77,7 +77,7 @@ def edit_election(self, request, form):
 
 
 @ElectionDayApp.form(model=Election, name='delete', template='form.pt',
-                     permission=Private, form=DeleteForm)
+                     permission=Private, form=EmptyForm)
 def delete_election(self, request, form):
 
     layout = ManageElectionsLayout(self, request)